[Dev] Generic receiver, part 6.
